new features

+++ Store.Models/Entities/Product.cs
@@ -1,10 +1,12 @@
+using System.ComponentModel.DataAnnotations.Schema;
+
 namespace Store.Models.Entities
 {
     public class Product
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double UnitValue { get; set; }
+        public long UnitValue { get; set; }
         public bool State { get; set; }
     }
 }
